requirements.txt
streamlit
openai
langchain
pymupdf
python-dotenv
faiss-cpu
tiktoken
langchain-community